CSV parser is now smart, it used attributes parseStringValue method to convert any input value to a correct internal value.
This should be implemented by e.g. dateattribute to convert any string
date to a valid date, by listattribute to select the correct value etc.

The analyzer immediately outputs the result of the parse, so the user
can see how the data is interpreted. Ignored columns are now more visually
ignored.


git-svn-id: http://svn.ibuildings.nl/atk/atk/trunk@1928 80284ba3-b555-44ee-8c06-7db2f0370df4